+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+### Added
+- Implement list command to list outputs and EDIDs
+
 ## [0.5.0] - 2016-06-05
 ### Added
 - EDID variable provided in `SRANDRD_ACTION`